+++ wagtail_localize_rws_languagecloud/forms.py
@@ -18,6 +18,62 @@ from wagtail_localize_rws_languagecloud.rws_client import ApiClient, NotAuthenti
 logger = logging.getLogger(__name__)
 
 
+# Initial values for the form
+# These are split out because they are also used elsewhere
+def get_default_project_template_id():
+    """Returns the default project template id for LanguageCloud projects"""
+    return settings.WAGTAILLOCALIZE_RWS_LANGUAGECLOUD.get("TEMPLATE_ID")
+
+
+def get_default_project_name_prefix():
+    """Returns the default project name prefix for LanguageCloud projects"""
+    prefix = settings.WAGTAILLOCALIZE_RWS_LANGUAGECLOUD.get("PROJECT_PREFIX", "")
+    return f"{prefix}{timezone.now():%Y-%m-%d}_"
+
+
+def get_default_due_date():
+    """Returns the default due date for LanguageCloud projects"""
+    now = timezone.now()
+    delta = settings.WAGTAILLOCALIZE_RWS_LANGUAGECLOUD.get(
+        "DUE_BY_DELTA", datetime.timedelta(days=7)
+    )
+    return now + delta
+
+
+def get_default_project_description(source_object, user=None):
+    """Get default description for the LanguageCloud project
+
+    Args:
+        source_object (TranslationSource): The source object
+        user (Optional[User]): The user who is creating the project
+
+    Returns:
+        str: The default description
+    """
+    description = ""
+    if user is not None:
+        description = user_display_name(user) + "\n"
+
+    if source_object is None:
+        return description
+
+    the_source_object = source_object
+    if isinstance(source_object, TranslationSource):
+        the_source_object = source_object.get_source_instance()
+
+    if isinstance(the_source_object, Page):
+        description = description + (the_source_object.full_url or "")
+        return description
+
+    pages = get_object_usage(the_source_object)
+    # This is only contextual information. If a snippet appears in hundreds of
+    # pages we probably don't need to spam all of them. Just take the first 5.
+    urls = [(page.full_url or "") for page in pages.all()[:5]]
+    description = description + "\n".join(urls)
+
+    return description
+
+
 class LanguageCloudProjectSettingsForm(WagtailAdminModelForm):
     class Meta:
         exclude = (
@@ -40,6 +96,7 @@ class LanguageCloudProjectSettingsForm(WagtailAdminModelForm):
         super().__init__(data, files, instance=instance, prefix=prefix, **kwargs)
 
         self.fields["user"].initial = user
+
         self.fields["user"].widget = forms.HiddenInput()
 
         self.fields["name"].label = _("Name prefix")
@@ -47,39 +104,22 @@ class LanguageCloudProjectSettingsForm(WagtailAdminModelForm):
             "The project name will be a combination of the "
             "supplied prefix and the source title. e.g. '%(name)s'"
             % {
-                "name": f"{self.default_project_name_prefix}{self._get_source_name(source_object_instance)}"
+                "name": f"{get_default_project_name_prefix()}{self._get_source_name(source_object_instance)}"
             }
         )
-        self.fields["name"].initial = self.default_project_name_prefix
-        self.fields["description"].initial = self._get_default_project_description(
+        self.fields["name"].initial = get_default_project_name_prefix()
+        self.fields["description"].initial = get_default_project_description(
             source_object_instance, user=user
         )
         self.fields["description"].widget = forms.Textarea(attrs={"rows": 3})
-        self.fields["due_date"].initial = self.default_due_date
+        self.fields["due_date"].initial = get_default_due_date()
 
         self.fields["template_id"] = forms.ChoiceField(
             label=_("Template"),
             choices=self._get_project_template_choices(),
-            initial=self.default_project_template_id,
+            initial=get_default_project_template_id(),
             widget=forms.Select(),
         )
-
-    @property
-    def default_project_template_id(self):
-        return settings.WAGTAILLOCALIZE_RWS_LANGUAGECLOUD.get("TEMPLATE_ID")
-
-    @property
-    def default_project_name_prefix(self):
-        prefix = settings.WAGTAILLOCALIZE_RWS_LANGUAGECLOUD.get("PROJECT_PREFIX", "")
-        return f"{prefix}{timezone.now():%Y-%m-%d}_"
-
-    @property
-    def default_due_date(self):
-        now = timezone.now()
-        delta = settings.WAGTAILLOCALIZE_RWS_LANGUAGECLOUD.get(
-            "DUE_BY_DELTA", datetime.timedelta(days=7)
-        )
-        return now + delta
 
     def clean_due_date(self):
         due_date = self.cleaned_data["due_date"]
@@ -96,30 +136,6 @@ class LanguageCloudProjectSettingsForm(WagtailAdminModelForm):
         else:
             source_name = str(source_object)
         return source_name
-
-    def _get_default_project_description(self, source_object, user=None):
-        description = ""
-        if user is not None:
-            description = user_display_name(user) + "\n"
-
-        if source_object is None:
-            return description
-
-        the_source_object = source_object
-        if isinstance(source_object, TranslationSource):
-            the_source_object = source_object.get_source_instance()
-
-        if isinstance(the_source_object, Page):
-            description = description + (the_source_object.full_url or "")
-            return description
-
-        pages = get_object_usage(the_source_object)
-        # This is only contextual information. If a snippet appears in hundreds of
-        # pages we probably don't need to spam all of them. Just take the first 5.
-        urls = [(page.full_url or "") for page in pages.all()[:5]]
-        description = description + "\n".join(urls)
-
-        return description
 
     def _get_project_templates(self):
         client = ApiClient(logger)
@@ -138,4 +154,4 @@ class LanguageCloudProjectSettingsForm(WagtailAdminModelForm):
                 for template in project_templates["items"]
             ]
 
-        return [(self.default_project_template_id, _("Default template"))]
+        return [(get_default_project_template_id(), _("Default template"))]