+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add new `update_translated_pages` management command
+
 ## [0.7.2] - 2022-03-24
 
 ### Added